+++ ext/etl/websocket_comm.go
@@ -14,7 +14,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/NVIDIA/aistore/api/apc"
 	"github.com/NVIDIA/aistore/cmn"
 	"github.com/NVIDIA/aistore/cmn/cos"
 	"github.com/NVIDIA/aistore/cmn/debug"
@@ -30,7 +29,7 @@ type (
 	// and creates per-xaction sessions that use those channels to exchange data with ETL pods.
 	statefulCommunicator interface {
 		Communicator
-		createSession(xctn core.Xact) (Session, error)
+		createSession(xctn core.Xact, multiplier int) (Session, error)
 	}
 
 	// Session represents a per-xaction communication context created by the statefulCommunicator.
@@ -41,43 +40,53 @@ type (
 		Finish(errCause error) error
 	}
 
-	// +--------------------------------------------------------------------------------------------+
-	// |                                       Target Node                                          |
-	// |            |                                                                               |
-	// |            |       +-------------------+                 +---------------------------+     |
-	// |            |       |   webSocketComm   | --- K8s API --> |   ETL Pod (Transformer)   |     |
-	// |            |       +---|-----------|---+                 +---------------------------+     |
-	// |            |           v           |                           ^ ^ ^ ^ ^   ^ ^ ^ ^ ^       |
-	// |            |       +-------------------+                       | | | | |   | | | | |       |
-	// |            |       |    wsSession (1) -|--> wsConnCtx (1)(1) <-+ | | | |   | | | | |       |
-	// |            |       |                / -|--> wsConnCtx (1)(2) <---+ | | |   | | | | |       |
-	// |  TCB (1) --------- | ---> workCh ---  -|--> wsConnCtx (1)(3) <-----+ | |   | | | | |       |
-	// |            |       |                \ -|--> wsConnCtx (1)(4) <-------+ |   | | | | |       |
-	// |            |       |                  -|--> wsConnCtx (1)(5) <---------+   | | | | |       |
-	// |            |       +-------------------+                                   | | | | |       |
-	// |            |                       |                                       | | | | |       |
-	// |            |                       v                           (ws conns)  | | | | |       |
-	// |            |       +-------------------+                                   | | | | |       |
-	// |            |       |    wsSession (2) -|--> wsConnCtx (2)(1) <-------------+ | | | |       |
-	// |            |       |                / -|--> wsConnCtx (2)(2) <---------------+ | | |       |
-	// |  TCB (2) --------- | ---> workCh ---  -|--> wsConnCtx (2)(3) <-----------------+ | |       |
-	// |            |       |                \ -| -> wsConnCtx (2)(4) <-------------------+-|       |
-	// |            |       |                  -| -> wsConnCtx (2)(5) <---------------------+       |
-	// |            |       +-------------------+                                                   |
-	// +--------------------------------------------------------------------------------------------+
+	//nolint:dupword // ASCII diagram contains repeated characters by design
+	/*
+	 * +--------------------------------------------------------------------------------------------+
+	 * |                                       Target Node                                          |
+	 * |            |                                                                               |
+	 * |            |       +-------------------+                  (#inlineSessionMultiplier)       |
+	 * |   inline   |       |    wsSession (0) -|--> wsConnCtx (0)(1) <--------------+              |
+	 * |  transform  ------ | ---> workCh ---  -|--> wsConnCtx (0)(2) <------------+ |              |
+	 * |  requests  |       |     (inline)     -|--> wsConnCtx (0)(3) <----------+ | |              |
+	 * |            |       +-------------------+                                | | |              |
+	 * |            |                  ^                                         v v v              |
+	 * |            |       +----------|--------+                 +---------------------------+     |
+	 * |            |       |   webSocketComm   | --- K8s API --> |   ETL Pod (Transformer)   |     |
+	 * |            |       +---|-----------|---+                 +---------------------------+     |
+	 * |            |           v           |                           ^ ^ ^ ^ ^   ^ ^ ^ ^ ^       |
+	 * |            |       +-------------------+                       | | | | |   | | | | |       |
+	 * |            |       |    wsSession (1) -|--> wsConnCtx (1)(1) <-+ | | | |   | | | | |       |
+	 * |            |       |                / -|--> wsConnCtx (1)(2) <---+ | | |   | | | | |       |
+	 * |  TCB (1) --------- | ---> workCh ---  -|--> wsConnCtx (1)(3) <-----+ | |   | | | | |       |
+	 * |            |       |                \ -|--> wsConnCtx (1)(4) <-------+ |   | | | | |       |
+	 * |            |       |    (offline)     -|--> wsConnCtx (1)(5) <---------+   | | | | |       |
+	 * |            |       +-------------------+                                   | | | | |       |
+	 * |            |                       |                                       | | | | |       |
+	 * |            |                       v                           (ws conns)  | | | | |       |
+	 * |            |       +-------------------+                                   | | | | |       |
+	 * |            |       |    wsSession (2) -|--> wsConnCtx (2)(1) <-------------+ | | | |       |
+	 * |            |       |                / -|--> wsConnCtx (2)(2) <---------------+ | | |       |
+	 * |  TCB (2) --------- | ---> workCh ---  -|--> wsConnCtx (2)(3) <-----------------+ | |       |
+	 * |            |       |                \ -| -> wsConnCtx (2)(4) <-------------------+-|       |
+	 * |            |       |    (offline)     -| -> wsConnCtx (2)(5) <---------------------+       |
+	 * |            |       +-------------------+                  (#offlineSessionMultiplier)      |
+	 * +--------------------------------------------------------------------------------------------+
+	 */
 
 	webSocketComm struct {
 		baseComm
-		commCtx       context.Context
-		sessions      map[string]Session
-		commCtxCancel context.CancelFunc
-		m             sync.Mutex
+		commCtx         context.Context
+		inlineSession   Session
+		offlineSessions map[string]Session
+		commCtxCancel   context.CancelFunc
+		m               sync.Mutex
 	}
 
 	wsSession struct {
 		txctn            core.Xact
 		sessionCtx       context.Context
-		workCh           chan transformTask
+		workCh           chan *transformTask
 		sessionCtxCancel context.CancelFunc
 		argType          string
 		connections      []*wsConnCtx
@@ -90,7 +99,7 @@ type (
 		txctn          core.Xact // tcb/tcobjs xaction that uses this session to perform transformatio
 		ctx            context.Context
 		conn           *websocket.Conn
-		workCh         chan transformTask  // outbound messages of the original objects to send to ETL pod
+		workCh         chan *transformTask // outbound messages of the original objects to send to ETL pod
 		writerCh       chan *io.PipeWriter // inbound (post-transform) messages from ETL pod
 		eg             *errgroup.Group
 		name           string
@@ -98,10 +107,15 @@ type (
 	}
 
 	transformTask struct {
-		r     io.ReadCloser
-		w     *io.PipeWriter
-		daddr string
-		fqn   string
+		Daddr  string `json:"dst_addr,omitempty"`
+		Targs  string `json:"etl_args,omitempty"`
+		FQN    string `json:"fqn,omitempty"`
+		rwpair `json:"-"`
+	}
+
+	rwpair struct {
+		r io.ReadCloser
+		w *io.PipeWriter
 	}
 )
 
@@ -116,6 +130,10 @@ const (
 	maxMsgSize = 16 * cos.GiB
 	// TODO: compare performance over different channel sizes
 	wockChSize = 512
+
+	// TODO -- FIXME: make these specific ETL configs `ws-multiplier`
+	inlineSessionMultiplier  = 1
+	offlineSessionMultiplier = 4
 )
 
 // SetupConnection establishes a test connection to the ETL pod's websocket endpoint.
@@ -125,82 +143,37 @@ func (ws *webSocketComm) SetupConnection() (err error) {
 		return err
 	}
 	ws.boot.uri += "/ws" // TODO: make this endpoint configurable
-	conn, resp, err := websocket.DefaultDialer.Dial(ws.boot.uri, nil)
-	if err != nil {
-		return err
-	}
-	resp.Body.Close()
-	return conn.Close() // closing after verifying connectivity
+
+	ws.inlineSession, err = ws.createSession(ws.boot.xctn, inlineSessionMultiplier)
+	return err
 }
 
-func (ws *webSocketComm) InlineTransform(w http.ResponseWriter, r *http.Request, lom *core.LOM, _ string) (int, error) {
-	// establish WebSocket connection to ETL
-	conn, resp, err := websocket.DefaultDialer.Dial(ws.boot.uri, nil)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-	defer resp.Body.Close()
-	defer conn.Close()
-
-	// if the ETL expects direct put, send an initial empty TextMessage before streaming the object content
-	if ws.boot.msg.IsDirectPut() {
-		conn.WriteMessage(websocket.TextMessage, []byte(""))
-	}
-
-	// send object content with configued arg type
-	writer, err := conn.NextWriter(websocket.BinaryMessage)
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
+func (ws *webSocketComm) InlineTransform(w http.ResponseWriter, _ *http.Request, lom *core.LOM, latestVer bool, targs string) (int, error) {
 	buf, slab := core.T.PageMM().Alloc()
 	defer slab.Free(buf)
 
-	switch ws.boot.msg.ArgType() {
-	case ArgTypeDefault, ArgTypeURL:
-		latestVer := r.URL.Query().Get(apc.QparamLatestVer)
-		srcResp := lom.GetROC(cos.IsParseBool(latestVer), false)
-		if srcResp.Err != nil {
-			return srcResp.Ecode, srcResp.Err
-		}
-		if _, err := cos.CopyBuffer(writer, srcResp.R, buf); err != nil {
-			return http.StatusInternalServerError, err
-		}
-		srcResp.R.Close()
-	case ArgTypeFQN:
-		if ecode, err := lomLoad(lom, ws.Xact().Kind()); err != nil {
-			return ecode, err
-		}
-		fqn := url.PathEscape(lom.FQN)
-		if _, err := writer.Write(cos.UnsafeB(fqn)); err != nil {
-			return http.StatusInternalServerError, err
-		}
+	resp := ws.inlineSession.OfflineTransform(lom, latestVer, false /*sync*/, &core.GetROCArgs{TransformArgs: targs})
+	if _, err := cos.CopyBuffer(w, resp.R, buf); err != nil {
+		return 0, err
 	}
-	writer.Close()
-
-	// receive transformed object back
-	ty, reader, err := conn.NextReader()
-	if err != nil {
-		return http.StatusInternalServerError, err
-	}
-	debug.Assert(ty == websocket.BinaryMessage)
-	if _, err := cos.CopyBuffer(w, reader, buf); err != nil {
-		return http.StatusInternalServerError, err
+	if resp.Err != nil {
+		return resp.Ecode, resp.Err
 	}
 
-	return 0, nil
+	return 0, resp.R.Close()
 }
 
-func (ws *webSocketComm) createSession(xctn core.Xact) (Session, error) {
+func (ws *webSocketComm) createSession(xctn core.Xact, multiplier int) (Session, error) {
 	if xctn == nil {
 		return nil, cos.NewErrNotFound(core.T, "invalid xact parameter")
 	}
 
-	connPerSession := ws.boot.config.TCB.SbundleMult * 4 // TODO: add specific ETL config on this
+	connPerSession := ws.boot.config.TCB.SbundleMult * multiplier // TODO: add specific ETL config on this
 	wss := &wsSession{
 		txctn:       xctn,
 		isDirectPut: ws.boot.msg.IsDirectPut(),
 		argType:     ws.boot.msg.ArgType(),
-		workCh:      make(chan transformTask, wockChSize),
+		workCh:      make(chan *transformTask, wockChSize),
 		connections: make([]*wsConnCtx, 0, connPerSession),
 	}
 	wss.sessionCtx, wss.sessionCtxCancel = context.WithCancel(ws.commCtx)
@@ -234,16 +207,17 @@ func (ws *webSocketComm) createSession(xctn core.Xact) (Session, error) {
 	}
 
 	ws.m.Lock()
-	ws.sessions[xctn.ID()] = wss
+	ws.offlineSessions[xctn.ID()] = wss
 	ws.m.Unlock()
 	return wss, nil
 }
 
 func (ws *webSocketComm) Stop() {
-	for _, session := range ws.sessions {
+	for _, session := range ws.offlineSessions {
 		session.Finish(cmn.ErrXactUserAbort)
 	}
-	ws.sessions = nil
+	ws.offlineSessions = nil
+	ws.inlineSession.Finish(cmn.ErrXactUserAbort)
 	ws.commCtxCancel()
 	ws.baseComm.Stop()
 }
@@ -254,11 +228,10 @@ func (ws *webSocketComm) Stop() {
 
 func (wss *wsSession) OfflineTransform(lom *core.LOM, latestVer, sync bool, gargs *core.GetROCArgs) core.ReadResp {
 	var (
-		daddr string
-		fqn   string
-		err   error
-		r     io.ReadCloser
-		oah   = &cos.SimpleOAH{Atime: time.Now().UnixNano()}
+		task = &transformTask{}
+		err  error
+		r    io.ReadCloser
+		oah  = &cos.SimpleOAH{Atime: time.Now().UnixNano()}
 	)
 	switch wss.argType {
 	case ArgTypeDefault, ArgTypeURL:
@@ -272,20 +245,28 @@ func (wss *wsSession) OfflineTransform(lom *core.LOM, latestVer, sync bool, garg
 		if ecode, err := lomLoad(lom, wss.txctn.Kind()); err != nil {
 			return core.ReadResp{Err: err, Ecode: ecode}
 		}
-		fqn = url.PathEscape(lom.FQN)
+		task.FQN = url.PathEscape(lom.FQN)
 	}
 
 	pr, pw := io.Pipe() // TODO -- FIXME: revise and remove
+	task.rwpair = rwpair{r, pw}
 
-	if wss.isDirectPut && gargs != nil && !gargs.Local {
-		daddr = gargs.Daddr
-		err = cmn.ErrSkip
+	if gargs != nil {
+		task.Targs = gargs.TransformArgs
+		if wss.isDirectPut && !gargs.Local && gargs.Daddr != "" {
+			task.Daddr = gargs.Daddr
+			err = cmn.ErrSkip
+		}
 	}
+
+	// local object should contain empty direct put address; remote object should contain valid direct put address
+	debug.Assert(gargs == nil || task.Daddr == "" && gargs.Local || task.Daddr != "" && !gargs.Local)
+	debug.Assert(task.Targs == "" || task.Daddr == "") // Targs is for inline transform, while Daddr is for offline transform
 
 	l, c := len(wss.workCh), cap(wss.workCh)
 	wss.chanFull.Check(l, c)
 
-	wss.workCh <- transformTask{r, pw, daddr, fqn}
+	wss.workCh <- task
 
 	return core.ReadResp{
 		R:     cos.NopOpener(pr),
@@ -332,7 +313,7 @@ func (wctx *wsConnCtx) readLoop() error {
 		default:
 			ty, r, err := wctx.conn.NextReader()
 			if err != nil {
-				err = fmt.Errorf("error on reading message from %s: %w", wctx.name, err)
+				err = fmt.Errorf("error reading message from %s: %w", wctx.name, err)
 				wctx.txctn.AddErr(err)
 				return err
 			}
@@ -346,7 +327,7 @@ func (wctx *wsConnCtx) readLoop() error {
 
 			if _, err = cos.CopyBuffer(writer, r, buf); err != nil {
 				cos.Close(writer)
-				err = fmt.Errorf("error on copying message from %s: %w", wctx.name, err)
+				err = fmt.Errorf("error copying message from %s: %w", wctx.name, err)
 				wctx.txctn.AddErr(err)
 				return err
 			}
@@ -367,7 +348,7 @@ func (wctx *wsConnCtx) writeLoop() error {
 			return nil
 		case <-wctx.etlxctn.ChanAbort():
 			return nil
-		case rw := <-wctx.workCh:
+		case task := <-wctx.workCh:
 			// Leverages the fact that WebSocket preserves message order and boundaries.
 			// Sends two consecutive WebSocket messages:
 			//   1. A TextMessage containing the direct PUT address
@@ -375,48 +356,36 @@ func (wctx *wsConnCtx) writeLoop() error {
 			//
 			// The ETL server is expected to consume them in the same order and treat them as logically linked.
 
-			// 1. send direct put address (empty string for local object)
-			err := wctx.conn.WriteMessage(websocket.TextMessage, cos.UnsafeB(rw.daddr))
+			// 1. send control message
+			err := wctx.conn.WriteMessage(websocket.BinaryMessage, cos.MustMarshal(*task))
 			if err != nil {
-				err = fmt.Errorf("error on writing direct put address %s: %w", wctx.name, err)
+				err = fmt.Errorf("error writing control message %s: %w", wctx.name, err)
 				wctx.txctn.AddErr(err)
 				return err
 			}
 
-			// 2. send object content or FQN path
-			switch {
-			case rw.fqn != "":
-				err := wctx.conn.WriteMessage(websocket.TextMessage, cos.UnsafeB(rw.fqn))
-				if err != nil {
-					err = fmt.Errorf("error on writing direct put address %s: %w", wctx.name, err)
-					wctx.txctn.AddErr(err)
-					return err
-				}
-			case rw.r != nil:
+			// 2. send object content if any (not fqn case)
+			if task.r != nil {
 				writer, err := wctx.conn.NextWriter(websocket.BinaryMessage)
 				if err != nil {
-					err = fmt.Errorf("error on getting next writter from %s: %w", wctx.name, err)
+					err = fmt.Errorf("error getting next writter from %s: %w", wctx.name, err)
 					wctx.txctn.AddErr(err)
 					return err
 				}
-				if _, err := cos.CopyBuffer(writer, rw.r, buf); err != nil {
-					cos.Close(rw.r)
-					cos.Close(writer)
-					err = fmt.Errorf("error on getting next writter from %s: %w", wctx.name, err)
+				if _, err := cos.CopyBuffer(writer, task.r, buf); err != nil {
+					err = fmt.Errorf("error getting next writter from %s: %w", wctx.name, err)
 					wctx.txctn.AddErr(err)
 					return err
 				}
-				cos.Close(rw.r)
+				cos.Close(task.r)
 				cos.Close(writer)
-			default:
-				debug.Assert(false, "transform task should be either reader or fqn")
 			}
 
 			// For each object sent to the ETL via `wctx.workCh`, we reserve a corresponding slot in `wctx.writerCh`
 			// for its writer. This guarantees that the correct writer is matched with the response when it arrives.
 			l, c := len(wctx.writerCh), cap(wctx.writerCh)
 			wctx.writerChanFull.Check(l, c)
-			wctx.writerCh <- rw.w
+			wctx.writerCh <- task.w
 		}
 	}
 }